+++ lib/src/transport/abstract_transport.dart
@@ -4,10 +4,23 @@ import '../message/abstract_message.dart';
 
 abstract class AbstractTransport {
   // make it possible to have a connection state in the transport
-  Completer onDisconnect = Completer();
+  Completer get onDisconnect;
+  Completer get onConnectionLost;
   Stream<AbstractMessage> receive();
-  Future<void> open();
-  Future<void> close();
+  Future<void> open({Duration pingInterval});
+  Future<void> close({error});
+  Future<void> get onReady;
   bool get isOpen;
+  bool get isReady;
   void send(AbstractMessage message);
+
+  void complete(_onDisconnect,error) {
+    if (_onDisconnect != null && !_onDisconnect.isCompleted){
+      if (error != null) {
+        _onDisconnect.complete();
+      } else {
+        _onDisconnect.complete(error);
+      }
+    }
+  }
 }